SpringTask1-Finished

+++ src/main/java/strategy/DiscountStrategy.java
@@ -7,5 +7,5 @@ import bean.User;
 import java.time.LocalDate;
 
 public interface DiscountStrategy {
-    int getDiscount(User user, EventAuditory eventAuditory);
+    double getDiscount(User user, EventAuditory eventAuditory);
 }
